Split test_ultima.py into ultima_tests
